Flights data

On-time data for all flights that departed NYC (i.e. JFK, LGA or EWR) in 2013.

Usage

rmr_t1

Format

Data frame with columns:

centre

Population centre name

year

Year of the observation

vacancy, turnover

Vacancy and turnover rates observed in October of year, as a porportion.

amr_2br

Average monthly rent for an in-sample 2-bedroom unit in October of yaer

amr_2br_fixed_lag_delta

For a fixed subset of units, proportional change in amr_2br between year - 1 to year

Source

CMHC Rental Market Survey Data Tables https://www.cmhc-schl.gc.ca/professionals/housing-markets-data-and-research/housing-data/data-tables/rental-market/rental-market-report-data-tables